Score the board in a terminal state.
The search relies on the best node having a high visit count in order to
pick it out. This iteracts badly with us not wanting to expand a node
that is terminal, and not updating the results in that case (thus not
recording a visit).

Make the search detect when it has reached a terminal node, count who
is the winner on the board, and initialize the SearchResult with that
information. This makes sure that winning positions can get a good visit
count.

Fixes issues #38 and #54.