Improve n^2 history copy behavior in long games.
Currently GameState contains a list of KoStates, each of which has the
full game history. As the game goes on, the search will keep getting slower
with each move played because it operates on GameState, and has to copy all
history data on each iteration. (It actually only needs the last 16
states)

As a workaround, instead of copying the KoStates, keep them as shared pointers
and reduce the work to copying the pointers and incrementing the refcounts.