Don't count playouts that did no evaluation.
If we try to expand a node that is already being expanded, for example
in deep endgames with little legal moves, where virtual loss can't keep
extra threads away from the same node, we will bail with SearchResult
set to invalid. This does not matter in normal gameplay because we will
try again, but when running with a fixed amount of playouts per move
this aborted expansion will still count towards the limit, causing the
program to play at a lower level than intended.

Check whether the traversal of the tree lead to any useful evaluation
(be it a node expansion or a final board count), and if not, don't count
the playout and try again.

Fixes issue #109.