Don't try to normalize noise if it is too small.
It's possible to generate a dirichtlet noise vector that sums to 0.0,
especially if the number of legal moves is very small. This then causes a
division by zero and all move scores to be NaN, hanging the engine.

Detect this case and bail out in the noise addition.

Probably fixes issue #117.