Change SPRT bounds to [0, 35].
After some investigation, [25, 35] bounds are too strict and require an
excessive amount of validation games to terminate. Switch to [0, 35].
This ensures it's very unlikely a weaker network is picked, makes it
very likely a 35 Elo improvement is not missed, and gives an expected
amount of validation games that is much more in line with the 400 from
the AlphaGo Zero paper.