Clean up thread count handling in options.
The benchmark changes hid the default threads setting from program
options in order to tell whether it was specified or not. Fix this by
changing the check to defaulted(), adding back the default setting, and
make the distiction between max threads and default threads by adding a
cfg_max_threads variable. This allows us to properly use the maximum
amount of threads in CPU-only mode.